+++ apps/misc/models.py
@@ -10,7 +10,3 @@ class Initiative(models.Model):
 
     def __str__(self):
         return self.title
-    
-          
-
-
