+++ templates/home/index.html
@@ -330,12 +330,4 @@
         }
     </script>
     <script>AOS.init({once: true});</script>
-{% endblock %}                            
-                                           
-                 
-                     
-                    
-                  
-                            
-                
-    
+{% endblock %}
